(#3) No exception on invalid user object
This may be used at times where a user object simply isn't
available and you still want to run is_demo(). If it throws
an exception you need to catch that, which makes the method
unfriendly to use.

It doesn't throw an exception anymore. Instead it will only
check the user object if its actually a User. Makes sense.